Update how the tip of the pen is found
Not searches for the biggest area of red, then finds the center of that. Then finds the center of every area of white, and compares that to the center of the largest red area. The tip of the pen is determined to be the center of white closest to the center of red.