BitMap only works for 35x35
So many magic values, idk why anything works. Nothing makes any sense. There are 51 bytes at the end that have no business being there, but if they aren't there, then the file won't open. Also, if I use windows to make a 35x35 bmp with 24-bit color depth, then it says that the raw data size is 3780, when it should really be 3729 (35 x 35 x 3 + 54).

Whatever, it works for what I need it to do, and I am sick of fooling with this. I need to move on.