Updated BitMap to set the image data using 2 12bit uint instead of 3 8bit unit
The method header allows for 16bits, but don't do that, it is a trap. Keep it under 4096. Doing this allows me to use higher accuracy x and y values. Previously, x-pos was stored in the blue channel, y-pos in green, and nothing in red. Now, x-pos is stored in blue and the top nibble of green, while y-pos is stored in bottom nibble of green and the blue channel. This brings the value range from 0-255 to 0-4096, a massive increase that should be able to handle a video feed of any resolution without any problems.